pcap support and restructuring (#4)
* Add all SBF blocks, v0.1

PVT confirmed to work from raw input data wrapper which wraps SBF blocks. Need to confirm all other sepentrio binary blocks

* Restructuring, v0.2

More restructuring to be done, serial and udp needs testing.

* Changes on driver for saving purposes

* Migrate helper utility into categorical message specification

Navigation, Calibration, Raw all have individual header files for readability

* checkpoint

* sbf pvt

* crc validation

* pvt geodetic

* Dispatcher for modularity

* Remove copyright

* restructure

* restructure publsiher

* revert back

* Restructure

* restructure v.01

* crash somewhere

* comment table

* add table checks

* logging

* typo

* endofpvt

* use global pvt

* not publishing pvt

* sync with hgkt

* Add PVT Cartesian

There is currently an issue with the incoming SBF data wrapped inside of InputDataWrapper. Mosaic uses 0x24 and 0x40 to sync, but messages can span across tow InputDataWrappers. This means that when the driver launches, if it is not aligned correctly with the bytestream, the chunks will not be decoded correctly. It is also the case s.t all of the SBF data uses the 0x24 and 0x40 sync keyword. Will look in the manual/septentrio github to see how they manage alignment

* change to debug

* pcap support

* playback functionality for debugging

* pr fixes

* remove status result

* include iomanip

* includes

* sbf covs

* print logs for what blocks are coming in

* print logs for what blocks are coming in

* cart v2

* Clean up driver, only use pvt_geodetic

* convert to degrees